move `tables_are_sorted()` test to `backend/tests.rs`
